try to get arm opt to work

+++ src/VecSim/spaces/functions/ARMPL_NEON.h
@@ -14,7 +14,7 @@ dist_func_t<float> Choose_FP32_IP_implementation_ARMPL_NEON(size_t dim);
 
 dist_func_t<float> Choose_FP32_L2_implementation_ARMPL_NEON(size_t dim);
 
-dist_func_t<float> Choose_FP64_IP_implementation_ARMPL_NEON(size_t dim);
+dist_func_t<double> Choose_FP64_IP_implementation_ARMPL_NEON(size_t dim);
 
-dist_func_t<float> Choose_FP64_L2_implementation_ARMPL_NEON(size_t dim);
+dist_func_t<double> Choose_FP64_L2_implementation_ARMPL_NEON(size_t dim);
 } // namespace spaces
